requirements.txt
ffmpeg-python
ultralytics
transformers
torch
torchvision
torchaudio
pillow
numpy
tqdm
scikit-learn